Move everything to `src`
